+++ src/hunter/_predicates.pyx
@@ -389,7 +389,9 @@ cdef class From:
         )
 
     def __repr__(self):
-        return '<hunter.predicates.From: condition=%r, predicate=%r>' % (self.condition, self.predicate)
+        return '<hunter.predicates.From: condition=%r, predicate=%r, watermark=%r>' % (
+            self.condition, self.predicate, self.watermark
+        )
 
     def __eq__(self, other):
         return (